Collector running per namespace
