+++ internal/nonkube/client/runtime/info.go
@@ -1,0 +1,35 @@
+package runtime
+
+import (
+	"fmt"
+
+	"github.com/skupperproject/skupper/internal/nonkube/client/fs"
+	"github.com/skupperproject/skupper/internal/nonkube/common"
+)
+
+func GetPlatform(namespace string) (string, error) {
+	platformLoader := &common.NamespacePlatformLoader{}
+	platform, err := platformLoader.Load(namespace)
+	if err != nil {
+		return "", err
+	}
+	return platform, err
+}
+
+func GetLocalRouterAddress(namespace string) (string, error) {
+	port, err := GetLocalRouterPort(namespace)
+	if err != nil {
+		return "", err
+	}
+	address := fmt.Sprintf("amqps://127.0.0.1:%d", port)
+	return address, nil
+}
+
+func GetLocalRouterPort(namespace string) (int, error) {
+	client := fs.NewRouterAccessHandler(namespace)
+	ra, err := client.Get("skupper-local")
+	if err != nil {
+		return 0, fmt.Errorf("unable to determine router port: %w", err)
+	}
+	return ra.Spec.Roles[0].Port, nil
+}
