+++ pkg/fs/watcher.go
@@ -2,7 +2,7 @@ package fs
 
 import (
 	"os"
-	"regexp"
+	"strings"
 	"time"
 
 	"github.com/fsnotify/fsnotify"
@@ -11,19 +11,21 @@ import (
 // FSChangeHandler provides a callback mechanism used by the FileWatcher
 // to notify about changes to monitored directory or file.
 type FSChangeHandler interface {
+	OnAdd(basePath string)
 	OnCreate(string)
 	OnUpdate(string)
 	OnRemove(string)
+	Filter(string) bool
 }
 
 // FileWatcher uses fsnotify to watch file system changes done to
 // files or directories, notifying the respective handlers. It is
 // recommended to watch directories over files (you can add filters
-// to limit the scope of files to be observed).
+// to limit the scope of files to be observed by your handler).
 type FileWatcher struct {
+	started    bool
 	watcher    *fsnotify.Watcher
-	handlerMap map[string]FSChangeHandler
-	filterMap  map[FSChangeHandler][]*regexp.Regexp
+	handlerMap map[string][]FSChangeHandler
 }
 
 func NewWatcher() (*FileWatcher, error) {
@@ -33,39 +35,50 @@ func NewWatcher() (*FileWatcher, error) {
 	}
 	return &FileWatcher{
 		watcher:    watcher,
-		handlerMap: map[string]FSChangeHandler{},
-		filterMap:  map[FSChangeHandler][]*regexp.Regexp{},
+		handlerMap: map[string][]FSChangeHandler{},
 	}, nil
 }
 
 func (w *FileWatcher) filterHandlers(name string) []FSChangeHandler {
-	var handlers []FSChangeHandler
+	var filteredHandlers []FSChangeHandler
 
-	for _, handler := range w.handlerMap {
-		if filters, ok := w.filterMap[handler]; !ok || len(filters) == 0 {
-			handlers = append(handlers, handler)
-		} else {
-			for _, f := range filters {
-				if f == nil {
-					continue
-				}
-				if f.MatchString(name) {
-					handlers = append(handlers, handler)
-					break
-				}
+	for baseName, handlers := range w.handlerMap {
+		if !strings.HasPrefix(name, baseName) {
+			continue
+		}
+		for _, handler := range handlers {
+			if handler.Filter(name) {
+				filteredHandlers = append(filteredHandlers, handler)
 			}
 		}
 	}
-	return handlers
+	return filteredHandlers
+}
+
+func (w *FileWatcher) prepareRemoved(event fsnotify.Event) {
+	if !event.Has(fsnotify.Remove) {
+		return
+	}
+	// When the specific event name being watched is removed,
+	// stop fsnotify and wait for it to show up again
+	if _, ok := w.handlerMap[event.Name]; ok {
+		_ = w.watcher.Remove(event.Name)
+		w.watchCreated(event.Name)
+	}
 }
 
 func (w *FileWatcher) Start(stopCh <-chan struct{}) {
+	if w.started {
+		return
+	}
+	w.started = true
 	go func() {
 		for {
 			select {
 			case event := <-w.watcher.Events:
 				handlers := w.filterHandlers(event.Name)
 				if len(handlers) == 0 {
+					w.prepareRemoved(event)
 					continue
 				}
 				switch {
@@ -81,11 +94,8 @@ func (w *FileWatcher) Start(stopCh <-chan struct{}) {
 					for _, handler := range handlers {
 						handler.OnRemove(event.Name)
 					}
-					// object being watched removed, watch for it to show up again
-					if handler, ok := w.handlerMap[event.Name]; ok {
-						_ = w.watcher.Remove(event.Name)
-						w.watchCreated(event.Name, handler)
-					}
+					// if object being watched is removed, watch for it to show up again
+					w.prepareRemoved(event)
 				}
 			case <-stopCh:
 				_ = w.watcher.Close()
@@ -99,14 +109,19 @@ func (w *FileWatcher) Start(stopCh <-chan struct{}) {
 // start watching the respective resource. It is recommended to
 // watch directories and filter the desired files, as watching
 // non-existing files directly might lead to missing events.
-func (w *FileWatcher) watchCreated(name string, handler FSChangeHandler) {
+func (w *FileWatcher) watchCreated(name string) {
 	go func() {
 		ticker := time.Tick(time.Second)
 		for {
 			select {
 			case <-ticker:
 				if err := w.watcher.Add(name); err == nil {
-					handler.OnCreate(name)
+					for _, handler := range w.handlerMap[name] {
+						handler.OnAdd(name)
+						if handler.Filter(name) {
+							handler.OnCreate(name)
+						}
+					}
 					return
 				}
 			}
@@ -115,12 +130,16 @@ func (w *FileWatcher) watchCreated(name string, handler FSChangeHandler) {
 	return
 }
 
-func (w *FileWatcher) Add(name string, handler FSChangeHandler, filters ...*regexp.Regexp) {
-	w.handlerMap[name] = handler
-	w.filterMap[handler] = filters
+func (w *FileWatcher) Add(name string, handler FSChangeHandler) {
+	handlers, ok := w.handlerMap[name]
+	if !ok {
+		w.handlerMap[name] = []FSChangeHandler{handler}
+	}
+	w.handlerMap[name] = append(handlers, handler)
 	if _, err := os.Stat(name); err != nil && os.IsNotExist(err) {
-		w.watchCreated(name, handler)
+		w.watchCreated(name)
 	} else {
 		_ = w.watcher.Add(name)
+		handler.OnAdd(name)
 	}
 }
