+++ pkg/vanflow/eventsource/manager.go
@@ -2,6 +2,7 @@ package eventsource
 
 import (
 	"context"
+	"encoding/json"
 	"errors"
 	"log/slog"
 	"time"
@@ -94,6 +95,8 @@ func (m *Manager) sendRecords(ctx context.Context) {
 			msg, err := record.Encode()
 			if err != nil {
 				m.logger.Error("skipping record message after encoding error:", slog.Any("error", err))
+				recordJson, _ := json.MarshalIndent(record, "", "  ")
+				m.logger.Error(string(recordJson))
 				continue
 			}
 			if err := sender.Send(ctx, msg); err != nil {
